[Refactor] handler에서 서비스 부분 분리, node 단위 API 요청으로 변경
